converse/requirements.txt
django==5.0.1
psycopg2-binary==2.9.6
gunicorn==21.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
binaryornot==0.4.4
blinker==1.8.2
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap4==2023.1
cryptography==42.0.2
cssselect2==0.7.0
Deprecated==1.2.14
distlib==0.3.8
django-braces==1.15.0
django-browser-reload==1.12.1
django-contrib-comments==2.2.0
django-contrib-comments==2.2.0
django-crispy-forms==2.1
django-etc==1.4.0
django-extensions==3.2.3
django-hitcount==1.3.5
django-htmx==1.17.3
django-model-utils==4.5.1
django-querysetsequence==0.17
django-taggit==6.0.0
django-tailwind==3.8.0
django-tinymce==3.7.1
docutils==0.20.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.1
Pillow==10.1.0
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyperclip==1.8.2
pyproject-api==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==4.3.4
reportlab==4.2.2
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.1
tox==4.17.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.19
virtualenv==20.26.3
webencodings==0.5.1
websockets==8.1
wrapt==1.16.0
